Update
More errors found in betting game addon. Patch
Further errors need to be reported

+++ lmo/addon/tipp/lmo-tippcalceinsicht.php
@@ -114,9 +114,9 @@ if (!file_exists($einsichtfile)) {
     $dum = explode('|', $tippdaten[$i-1]);
     $op1 = $dum[0];
     // Nick
-    $op3 = substr($dum[1], 2)-1;
+    $op3 = (int)substr($dum[1], 2)-1;
     // Spieltagsnummer
-    $op4 = substr($dum[1], 0, 2);
+    $op4 = (int)substr($dum[1], 0, 2);
     // TP
     $op6 = (int)substr($dum[1], 2, -1)-1;
     $op7 = (int)substr($dum[1], -1)-1;
